// src/job-offers/interfaces/provider1-response.interface.ts
export interface Provider1Response {
    metadata: {
        requestId: string;
        timestamp: string;
    };
    jobs: Provider1Job[];
}

export interface Provider1Job {
    jobId: string;
    title: string;
    details: {
        location: string;
        type: string;
        salaryRange: string;
    };
    company: {
        name: string;
        industry: string;
    };
    skills: string[];
    postedDate: string;
}